+++ src/main/java/project/linkarchive/backend/url/controller/UrlQueryController.java
@@ -1,4 +1,60 @@
 package project.linkarchive.backend.url.controller;
 
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import project.linkarchive.backend.url.response.UrlMetaDataResponse;
+
+import java.io.IOException;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+
+@RestController
 public class UrlQueryController {
+
+    @GetMapping("/url/metadata")
+    public ResponseEntity<UrlMetaDataResponse> getUrlMetaData(@RequestParam(value = "url") String url) {
+        URL urlObject;
+        try {
+            urlObject = new URL(url);
+        } catch (MalformedURLException e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
+        HttpURLConnection connection;
+        try {
+            connection = (HttpURLConnection) urlObject.openConnection();
+            connection.setRequestMethod("GET");
+            connection.connect();
+            int responseCode = connection.getResponseCode();
+            if (responseCode != HttpURLConnection.HTTP_OK) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
+
+        Document document;
+        try {
+            document = Jsoup.connect(url).get();
+        } catch (IOException e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
+        }
+
+        String title = document.select("meta[property=og:title]").attr("content");
+        String description = document.select("meta[property=og:description]").attr("content");
+        String thumbnail = document.select("meta[property=og:image]").attr("content");
+
+        UrlMetaDataResponse urlMetaDataResponse = new UrlMetaDataResponse(title, description, thumbnail);
+        return ResponseEntity.ok(urlMetaDataResponse);
+    }
+
 }
